+++ lib/src/theme/tab_theme_data.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/widgets.dart';
-import 'package:tabbed_view/src/icon_path.dart';
+import 'package:tabbed_view/src/icon_provider.dart';
 import 'package:tabbed_view/src/tabbed_view_icons.dart';
 import 'package:tabbed_view/src/theme/button_colors.dart';
 import 'package:tabbed_view/src/theme/tab_status_theme_data.dart';
@@ -9,8 +9,7 @@ import 'package:tabbed_view/src/theme/vertical_alignment.dart';
 /// Theme for tab.
 class TabThemeData {
   TabThemeData(
-      {IconData? closeIconData,
-      IconPath? closeIconPath,
+      {IconProvider? closeIcon,
       this.buttonColors = const ButtonColors(),
       double buttonIconSize = TabbedViewThemeConstants.defaultIconSize,
       this.verticalAlignment = VerticalAlignment.center,
@@ -30,10 +29,9 @@ class TabThemeData {
         this._buttonsGap = buttonsGap >= 0 ? buttonsGap : 0,
         this.buttonIconSize =
             TabbedViewThemeConstants.normalize(buttonIconSize),
-        this.closeIconData = closeIconData,
-        this.closeIconPath = (closeIconData == null && closeIconPath == null)
-            ? TabbedViewIcons.close
-            : closeIconPath,
+        this.closeIcon = closeIcon == null
+            ? IconProvider.path(TabbedViewIcons.close)
+            : closeIcon,
         this.selectedStatus =
             selectedStatus != null ? selectedStatus : TabStatusThemeData(),
         this.highlightedStatus = highlightedStatus != null
@@ -64,11 +62,8 @@ class TabThemeData {
   ButtonColors buttonColors;
   double _buttonsOffset;
 
-  /// Icon data for the close button.
-  IconData? closeIconData;
-
-  /// Icon path for the close button.
-  IconPath? closeIconPath;
+  /// Icon for the close button.
+  IconProvider closeIcon;
 
   TabStatusThemeData selectedStatus;
   TabStatusThemeData highlightedStatus;